+++ webots_ros2_core/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package webots_ros2_core
 1.0.0 (2019-XX-YY)
 ------------------
 * Added support for multi robots.
+* Added a new tf_publisher node to publish tf of desired Webots nodes.
 
 0.0.2 (2019-09-23)
 ------------------